+++ R/wapor_res_key.R
@@ -74,12 +74,14 @@
 #'
 #' @keywords internal
 #' @examples
+#' \dontrun{
 #' wapor_res_key("L1-AETI-D")   # "L1_300m"
 #' wapor_res_key("L1-PCP-D")    # "L1_5000m"
 #' wapor_res_key("L1-RET-D")    # "L1_30000m"
 #' wapor_res_key("L2-AETI-D")   # "L2_100m"
 #' wapor_res_key("L3-AETI-D")   # "L3_30m"
 #' wapor_res_key("AGERA5-ET0-E") # "AGERA5_11000m"
+#' }
 wapor_res_key <- function(variable) {
   stopifnot(is.character(variable), length(variable) == 1L)